+++ src/main/java/org/jenkinsci/plugins/ewm/strategies/FastestWriteSpeedStrategy.java
@@ -29,7 +29,7 @@ public class FastestWriteSpeedStrategy extends AbstractDiskSpeedStrategy {
     }
 
     @Override
-    protected double getDiskSpeed(@Nonnull DiskInfoProvider diskInfo) {
+    protected int getDiskSpeed(@Nonnull DiskInfoProvider diskInfo) {
         return diskInfo.getWriteSpeed();
     }
